pyproject.toml
[tool.poetry]
name = "bookish"
version = "0.1.0"
description = ""
authors = ["Tom Johnson <tom.johnson@softwire.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.2"
psycopg2 = "^2.9.3"
Flask-SQLAlchemy = "^3.0.0"
Flask-Script = "^2.0.6"
Flask-Migrate = "^3.1.0"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.20.0"
greenlet = "==3.0.3"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
